`jovyan` user needs to use the UID/GID of the logged in user
`jovyan` user assumes UID: 1000 and GID: 1000 as of now. However, that
will not work (as seen with one of the customer working sessions) when
the logged in user is not using 1000:1000 and downloads fiddler-samples,
since the files will be stored using the UID:GID of the logged in user.

The best practice here is for `jovyan` user to assume the UID:GID of the
logged in user.